Extract streaming pipeline coordinators
