+++ fiberhmm/inference/parallel.py
@@ -5,8 +5,7 @@ import shutil
 import sys
 import tempfile
 import time
-from collections import deque
-from concurrent.futures import Future, ProcessPoolExecutor, as_completed
+from concurrent.futures import ProcessPoolExecutor, as_completed
 from typing import Optional, Set, Tuple
 
 import numpy as np
@@ -20,10 +19,12 @@ from fiberhmm.inference.bam_output import (
 from fiberhmm.inference.engine import (
     _extract_fiber_read_from_pysam,
     _process_single_read,
-    make_apply_payload,
 )
 from fiberhmm.inference.mp_context import _MP_CONTEXT, _select_mp_context  # noqa: F401
-from fiberhmm.inference.read_filters import ReadFilterConfig, streaming_skip_reason
+from fiberhmm.inference.read_filters import (  # noqa: F401
+    ReadFilterConfig,
+    streaming_skip_reason,
+)
 from fiberhmm.inference.region_planning import (
     _get_genome_regions,
 )
@@ -46,15 +47,19 @@ from fiberhmm.inference.region_workers import (
     _process_region_to_bed,
 )
 from fiberhmm.inference.streaming_drain import (
-    _drain_oldest_chunk,
-    _drain_oldest_fused_chunk,
+    _drain_oldest_chunk,  # noqa: F401
+    _drain_oldest_fused_chunk,  # noqa: F401
+)
+from fiberhmm.inference.streaming_pipeline import (
+    _process_bam_streaming_pipeline,
+    _process_bam_streaming_pipeline_fused,  # noqa: F401
 )
 from fiberhmm.inference.streaming_workers import (
     _init_bam_worker,
-    _init_fused_worker,
+    _init_fused_worker,  # noqa: F401
     _process_chunk_worker,
-    _process_fused_payload_chunk_worker,
-    _process_payload_chunk_worker,
+    _process_fused_payload_chunk_worker,  # noqa: F401
+    _process_payload_chunk_worker,  # noqa: F401
 )
 from fiberhmm.inference.tagging import (
     set_legacy_apply_tags,
@@ -605,486 +610,6 @@ def _process_bam_region_parallel_fused(
         shutil.rmtree(temp_dir, ignore_errors=True)
 
     return aggregation.total_reads, aggregation.reads_with_footprints
-
-
-def _process_bam_streaming_pipeline_fused(
-    input_bam: str, output_bam: str,
-    model_path: str, recall_model_path: str,
-    train_rids,
-    edge_trim: int, circular: bool,
-    mode: str, context_size: int,
-    msp_min_size: int, nuc_min_size: int,
-    min_mapq: int, prob_threshold: int, min_read_length: int,
-    with_scores: bool,
-    min_llr: float, min_opps: int, unify_threshold: int,
-    emission_uplift: float,
-    also_write_legacy: bool, downstream_compat: bool,
-    max_reads: int, n_cores: int, chunk_size: int,
-    io_threads: int,
-    process_unmapped: bool = False,
-    primary_only: bool = False,
-    ref_fasta_path: Optional[str] = None,
-):
-    """Fused apply+recall streaming pipeline: one worker pool does both
-    the HMM nucleosome/MSP calls and the TF Kadane scan per read.
-
-    Eliminates the apply→recall pipe serialization (which was
-    ~2-7× slower than the sum of the two stages run alone on this Mac
-    benchmark) and the duplicate MM/ML parse + encode done by the
-    streaming recall stage.
-    """
-    # Streaming mode reads the BAM sequentially in the main process, so it can
-    # hand a live FastaFile to make_apply_payload for on-the-fly MD fallback on
-    # raw DAF BAMs. The handle is opened after BAM output setup and closed from
-    # the processing finally block so failures do not leak it.
-    ref_fasta = None
-    pysam.set_verbosity(0)
-    max_inflight = n_cores + 2
-    start_time = time.time()
-    counters = {
-        'reads_with_footprints': 0,
-        'no_footprints': 0,
-        'worker_failures': 0,
-        'written': 0,
-    }
-
-    total_reads = 0
-    skipped = 0
-    skip_reasons = {
-        'unmapped': 0, 'secondary_supplementary': 0, 'low_mapq': 0,
-        'too_short': 0, 'training_excluded': 0, 'no_modifications': 0,
-        'extraction_failed': 0,
-    }
-    filter_config = ReadFilterConfig(
-        min_mapq=min_mapq,
-        min_read_length=min_read_length,
-        primary_only=primary_only,
-        process_unmapped=process_unmapped,
-        train_rids=train_rids,
-    )
-
-    _log = sys.stderr if output_bam == '-' else sys.stdout
-
-    _output_target = output_bam
-    if output_bam == '-':
-        _output_target = os.fdopen(1, 'wb', closefd=False)
-
-    with pysam.AlignmentFile(input_bam, "rb", threads=io_threads, check_sq=False) as inbam:
-        with pysam.AlignmentFile(_output_target, "wb", header=inbam.header, threads=io_threads) as outbam:
-            if ref_fasta_path:
-                ref_fasta = pysam.FastaFile(ref_fasta_path)
-
-            executor = ProcessPoolExecutor(
-                max_workers=n_cores,
-                mp_context=_MP_CONTEXT,
-                initializer=_init_fused_worker,
-                initargs=(model_path, recall_model_path, emission_uplift, False),
-            )
-
-            inflight = deque()
-            chunk_payloads = []
-            chunk_read_objs = []
-            chunk_skip_flags = []
-            last_progress_reads = 0
-            last_progress_time = time.time()
-
-            def _buffer_skip(rd, reason):
-                nonlocal skipped
-                chunk_read_objs.append(rd)
-                chunk_skip_flags.append(True)
-                skipped += 1
-                skip_reasons[reason] += 1
-
-            try:
-                for read in inbam.fetch(until_eof=True):
-                    skip_reason = streaming_skip_reason(read, filter_config)
-                    if skip_reason:
-                        _buffer_skip(read, skip_reason)
-                        continue
-
-                    payload = make_apply_payload(read, mode=mode, ref_fasta=ref_fasta)
-                    if payload is None:
-                        _buffer_skip(read, 'no_modifications')
-                        continue
-
-                    chunk_payloads.append(payload)
-                    chunk_read_objs.append(read)
-                    chunk_skip_flags.append(False)
-                    total_reads += 1
-
-                    if max_reads and total_reads >= max_reads:
-                        break
-
-                    if len(chunk_read_objs) >= chunk_size:
-                        if len(inflight) >= max_inflight:
-                            _drain_oldest_fused_chunk(
-                                inflight, outbam, with_scores,
-                                also_write_legacy, downstream_compat, counters,
-                            )
-                        future = executor.submit(
-                            _process_fused_payload_chunk_worker,
-                            chunk_payloads, edge_trim, circular, mode,
-                            context_size, msp_min_size, nuc_min_size,
-                            with_scores, prob_threshold,
-                            mode, context_size,
-                            min_llr, min_opps, unify_threshold,
-                        )
-                        inflight.append((future, chunk_read_objs, chunk_payloads, chunk_skip_flags))
-                        chunk_payloads = []
-                        chunk_read_objs = []
-                        chunk_skip_flags = []
-
-                        now = time.time()
-                        elapsed = now - start_time
-                        avg = total_reads / elapsed if elapsed > 0 else 0
-                        dt = now - last_progress_time
-                        inst = (total_reads - last_progress_reads) / dt if dt > 0 else 0
-                        last_progress_reads = total_reads
-                        last_progress_time = now
-                        print(f"\r  Fused: {total_reads:,} | Skipped: {skipped:,} | "
-                              f"Inflight: {len(inflight)} | {inst:.0f} r/s (avg {avg:.0f})",
-                              end='', file=_log)
-                        _log.flush()
-
-                # Final partial chunk
-                if chunk_read_objs:
-                    if len(inflight) >= max_inflight:
-                        _drain_oldest_fused_chunk(
-                            inflight, outbam, with_scores,
-                            also_write_legacy, downstream_compat, counters,
-                        )
-                    if chunk_payloads:
-                        future = executor.submit(
-                            _process_fused_payload_chunk_worker,
-                            chunk_payloads, edge_trim, circular, mode,
-                            context_size, msp_min_size, nuc_min_size,
-                            with_scores, prob_threshold,
-                            mode, context_size,
-                            min_llr, min_opps, unify_threshold,
-                        )
-                    else:
-                        future = Future()
-                        future.set_result([])
-                    inflight.append((future, chunk_read_objs, chunk_payloads, chunk_skip_flags))
-
-                while inflight:
-                    _drain_oldest_fused_chunk(
-                        inflight, outbam, with_scores,
-                        also_write_legacy, downstream_compat, counters,
-                    )
-            finally:
-                try:
-                    executor.shutdown(wait=True)
-                finally:
-                    if ref_fasta is not None:
-                        ref_fasta.close()
-                        ref_fasta = None
-
-    elapsed = time.time() - start_time
-    rate = total_reads / elapsed if elapsed > 0 else 0
-    reads_with_fp = counters['reads_with_footprints']
-    print(f"\r  Fused: {total_reads:,} | Skipped: {skipped:,} | "
-          f"With footprints: {reads_with_fp:,} | {rate:.1f} r/s", file=_log)
-    if counters['worker_failures']:
-        print(
-            f"  Worker read failures: {counters['worker_failures']:,} "
-            f"(passed through unchanged)",
-            file=_log,
-        )
-    if skipped > 0:
-        print("  Skip reasons:", file=_log)
-        for reason, count in sorted(skip_reasons.items(), key=lambda x: -x[1]):
-            if count > 0:
-                pct = 100 * count / (total_reads + skipped)
-                print(f"    {reason}: {count:,} ({pct:.1f}%)", file=_log)
-
-    return total_reads, reads_with_fp
-
-
-def _process_bam_streaming_pipeline(
-    input_bam: str, output_bam: str,
-    model_path: str, train_rids: Set[str],
-    edge_trim: int, circular: bool,
-    mode: str, context_size: int,
-    msp_min_size: int,
-    nuc_min_size: int = 85,
-    min_mapq: int = 0,
-    prob_threshold: int = 0,
-    min_read_length: int = 0,
-    with_scores: bool = False,
-    n_cores: int = 4,
-    chunk_size: int = 500,
-    max_inflight: Optional[int] = None,
-    io_threads: int = 4,
-    primary_only: bool = False,
-    output_posteriors: Optional[str] = None,
-    write_msps: bool = True,
-    max_reads: Optional[int] = None,
-    debug_timing: bool = False,
-    process_unmapped: bool = False,
-) -> Tuple[int, int]:
-    """
-    Streaming producer-consumer pipeline for BAM processing.
-
-    Note: ``ref_fasta`` for the DAF MD-only fallback is set to None here;
-    --reference CLI plumbing is Step 4 of the daf-md fallback work.
-    MD-only DAF BAMs work without it.
-
-    Uses a sliding window of in-flight futures to overlap I/O and compute.
-    Works with unaligned/unindexed BAMs and stdin. Order-preserving.
-
-    The main process reads BAM sequentially, extracts fiber data, and submits
-    chunks to a process pool. Multiple chunks can be in flight simultaneously
-    (bounded by max_inflight), keeping workers saturated while the main
-    process continues reading and writing.
-
-    Args:
-        max_inflight: Max chunks in flight (default: 2 * n_cores)
-        chunk_size: Reads per compute chunk (default: 500)
-        io_threads: htslib decompression threads (default: 4)
-        process_unmapped: If True, process unmapped reads that have sequences
-
-    Returns:
-        (total_reads_processed, reads_with_footprints)
-    """
-    if max_inflight is None:
-        max_inflight = 2 * n_cores
-
-    pysam.set_verbosity(0)
-
-    ref_fasta = None    # placeholder; --reference plumbing is Step 4 of the
-                        # daf-md fallback work. MD-only DAF BAMs work without it.
-
-    total_reads = 0
-    skip_reasons = {
-        'unmapped': 0,
-        'secondary_supplementary': 0,
-        'low_mapq': 0,
-        'too_short': 0,
-        'training_excluded': 0,
-        'no_modifications': 0,
-        'extraction_failed': 0,
-        'no_footprints': 0,
-    }
-    filter_config = ReadFilterConfig(
-        min_mapq=min_mapq,
-        min_read_length=min_read_length,
-        primary_only=primary_only,
-        process_unmapped=process_unmapped,
-        train_rids=train_rids,
-    )
-
-    counters = {
-        'reads_with_footprints': 0,
-        'no_footprints': 0,
-        'worker_failures': 0,
-        'written': 0,
-    }
-
-    # Initialized after BAM handles open so failure cleanup is centralized in
-    # the processing finally block.
-    posterior_writer = None
-    posterior_stats = None
-    return_posteriors = False
-
-    # When writing to stdout, send progress to stderr to avoid corrupting BAM
-    _log = sys.stderr if output_bam == '-' else sys.stdout
-
-    print(f"Processing BAM (streaming pipeline, {n_cores} workers, "
-          f"chunk_size={chunk_size}, max_inflight={max_inflight})...", file=_log)
-    _log.flush()
-
-    start_time = time.time()
-
-    # For stdout output, ensure pysam writes to the real stdout fd
-    # (sys.stdout may have been redirected to stderr for logging)
-    _output_target = output_bam
-    if output_bam == '-':
-        _output_target = os.fdopen(1, 'wb', closefd=False)
-
-    with pysam.AlignmentFile(input_bam, "rb", threads=io_threads, check_sq=False) as inbam:
-        with pysam.AlignmentFile(_output_target, "wb", header=inbam.header, threads=io_threads) as outbam:
-
-            if output_posteriors:
-                if HAS_POSTERIOR_WRITER:
-                    posterior_writer = PosteriorWriter(
-                        output_posteriors, mode, context_size,
-                        edge_trim, input_bam, batch_size=1000
-                    )
-                    return_posteriors = True
-                    print(f"Posteriors will be written to: {output_posteriors}", file=_log)
-                else:
-                    print(
-                        "WARNING: posterior_writer.py not found, skipping posteriors export",
-                        file=_log,
-                    )
-
-            # Create worker pool
-            executor = ProcessPoolExecutor(
-                max_workers=n_cores,
-                mp_context=_MP_CONTEXT,
-                initializer=_init_bam_worker,
-                initargs=(model_path, debug_timing)
-            )
-
-            inflight = deque()
-            chunk_reads = []           # only fiber_reads to be processed
-            chunk_read_objs = []       # ALL reads in stream order (skipped + processed)
-            chunk_skip_flags = []      # bool per chunk_read_objs entry: True = pass-through
-            skipped = 0
-            last_progress_reads = 0
-            last_progress_time = time.time()
-
-            def _buffer_skip(rd, reason):
-                """Buffer a skipped read at its stream position to preserve sort order."""
-                nonlocal skipped
-                chunk_read_objs.append(rd)
-                chunk_skip_flags.append(True)
-                skipped += 1
-                skip_reasons[reason] += 1
-
-            try:
-                for read in inbam.fetch(until_eof=True):
-                    # --- Filter phase ---
-                    # Skipped reads are buffered into the chunk (with skip_flag=True)
-                    # rather than written immediately — the drain step writes them
-                    # at their original stream position so a coordinate-sorted
-                    # input BAM produces a coordinate-sorted output BAM.
-                    skip_reason = streaming_skip_reason(read, filter_config)
-                    if skip_reason:
-                        _buffer_skip(read, skip_reason)
-                        continue
-
-                    # --- Build slim payload (fast — no MM/ML parsing) ---
-                    # The MM/ML parse is now done inside the worker (see
-                    # _process_payload_chunk_worker) which lifts the apply
-                    # throughput ceiling from ~150-300 r/s (serial main parse
-                    # bottleneck) to whatever the worker pool can sustain.
-                    # Reads with no usable modification data or per-read worker
-                    # failures come back as result=None and are written through
-                    # without footprint tags (handled in _drain_oldest_chunk).
-                    payload = make_apply_payload(read, mode=mode, ref_fasta=ref_fasta)
-                    if payload is None:
-                        _buffer_skip(read, 'no_modifications')
-                        continue
-
-                    # --- Buffer into chunk ---
-                    chunk_reads.append(payload)
-                    chunk_read_objs.append(read)
-                    chunk_skip_flags.append(False)
-                    total_reads += 1
-
-                    # Check max reads limit
-                    if max_reads and total_reads >= max_reads:
-                        break
-
-                    # --- Submit when chunk is full ---
-                    # Use chunk_read_objs length so skipped reads also count toward
-                    # the chunk size (otherwise pure-skip stretches would buffer
-                    # unboundedly waiting for processable reads).
-                    if len(chunk_read_objs) >= chunk_size:
-                        # Drain oldest if at capacity
-                        if len(inflight) >= max_inflight:
-                            _drain_oldest_chunk(
-                                inflight, outbam, with_scores, write_msps,
-                                posterior_writer, counters
-                            )
-
-                        # Submit chunk of slim payloads to slim-IPC workers
-                        future = executor.submit(
-                            _process_payload_chunk_worker,
-                            chunk_reads, edge_trim, circular, mode, context_size,
-                            msp_min_size, nuc_min_size, with_scores, return_posteriors,
-                            prob_threshold,
-                        )
-                        inflight.append((future, chunk_read_objs, chunk_reads, chunk_skip_flags))
-                        chunk_reads = []
-                        chunk_read_objs = []
-                        chunk_skip_flags = []
-
-                        # Progress update
-                        now = time.time()
-                        elapsed = now - start_time
-                        avg_rate = total_reads / elapsed if elapsed > 0 else 0
-                        dt = now - last_progress_time
-                        inst_rate = (total_reads - last_progress_reads) / dt if dt > 0 else 0
-                        last_progress_reads = total_reads
-                        last_progress_time = now
-                        print(f"\r  Processed: {total_reads:,} | "
-                              f"Skipped: {skipped:,} | "
-                              f"Inflight: {len(inflight)} | "
-                              f"{inst_rate:.0f} reads/s (avg {avg_rate:.0f})", end='', file=_log)
-                        _log.flush()
-
-                # Submit final partial chunk (if it has any reads — skipped or not).
-                # Pass-through-only tail (e.g. unmapped reads at end) still needs
-                # to be flushed in order through the drain mechanism.
-                if chunk_read_objs:
-                    if len(inflight) >= max_inflight:
-                        _drain_oldest_chunk(
-                            inflight, outbam, with_scores, write_msps,
-                            posterior_writer, counters
-                        )
-
-                    if chunk_reads:
-                        future = executor.submit(
-                            _process_payload_chunk_worker,
-                            chunk_reads, edge_trim, circular, mode, context_size,
-                            msp_min_size, nuc_min_size, with_scores, return_posteriors,
-                            prob_threshold,
-                        )
-                    else:
-                        # All-skipped tail: no worker call needed, fake a
-                        # completed future returning [].
-                        future = Future()
-                        future.set_result([])
-                    inflight.append((future, chunk_read_objs, chunk_reads, chunk_skip_flags))
-
-                # Drain all remaining in-flight chunks (in order)
-                while inflight:
-                    _drain_oldest_chunk(
-                        inflight, outbam, with_scores, write_msps,
-                        posterior_writer, counters
-                    )
-
-            finally:
-                try:
-                    executor.shutdown(wait=True)
-                finally:
-                    if posterior_writer:
-                        posterior_stats = posterior_writer.close()
-                        posterior_writer = None
-
-    elapsed = time.time() - start_time
-    rate = total_reads / elapsed if elapsed > 0 else 0
-    reads_with_footprints = counters['reads_with_footprints']
-    print(f"\r  Processed: {total_reads:,} | Skipped: {skipped:,} | "
-          f"With footprints: {reads_with_footprints:,} | {rate:.1f} reads/s", file=_log)
-    if counters['worker_failures']:
-        print(
-            f"  Worker read failures: {counters['worker_failures']:,} "
-            f"(passed through unchanged)",
-            file=_log,
-        )
-
-    # Print skip reasons
-    if skipped > 0:
-        print("  Skip reasons:", file=_log)
-        for reason, count in sorted(skip_reasons.items(), key=lambda x: -x[1]):
-            if count > 0:
-                pct = 100 * count / (total_reads + skipped)
-                print(f"    {reason}: {count:,} ({pct:.1f}%)", file=_log)
-
-    # Sort and index output BAM (skip for stdout and unaligned mode)
-    if output_bam != '-' and not process_unmapped:
-        _sort_and_index_bam(output_bam, threads=n_cores)
-
-    # Report posteriors after the writer has been closed by the processing finally block.
-    if posterior_stats:
-        n_fibers, file_size = posterior_stats
-        print(f"Posteriors: {n_fibers:,} fibers -> {output_posteriors} ({file_size:.1f} MB)", file=_log)
-
-    return total_reads, reads_with_footprints
 
 
 def process_bam_for_footprints(input_bam: str, output_bam: str,
